ma2.0.0 - App, `kart` now is added as a module!
The following files were edited by adding the prefix
`'.' + MODULE_PATH` to every url.

     modified:   app/kart/index.html
     modified:   app/kart/kart.html
     modified:   app/kart/pepa.py
     modified:   app/kart/static/js/a.js
     modified:   app/kart/static/js/index.js

For APIs in `js` files, a global variable is defined in the
corresponding `html` with the module's path.

This release standardizes my own way of building bad practiced
modules. The approach is simple and aims to building modules which
need to work as standalone apps when executed from their own folders
and behave the same as if they were started as a module from a parent
folder.